Rename basePickerStyles to pickerStyles
