--- src/components/Picker/BasePicker/basePickerStyles/index.android.js
@@ -1,11 +1,0 @@
-import styles from '../../../../styles/styles';
-
-const pickerStyles = disabled => ({
-    ...styles.picker(disabled),
-    inputAndroid: {
-        ...styles.picker(disabled).inputNative,
-        paddingLeft: 12,
-    },
-});
-
-export default pickerStyles;
