--- src/components/Picker/BasePicker/basePickerStyles/index.ios.js
@@ -1,8 +1,0 @@
-import styles from '../../../../styles/styles';
-
-const pickerStyles = disabled => ({
-    ...styles.picker(disabled),
-    inputIOS: styles.picker(disabled).inputNative,
-});
-
-export default pickerStyles;
