+++ src/components/Picker/pickerStyles/index.ios.js
@@ -1,0 +1,9 @@
+// eslint-disable-next-line rulesdir/display-name-property
+import styles from '../../../styles/styles';
+
+const pickerStyles = isDisabled => ({
+    ...styles.picker(isDisabled),
+    inputIOS: styles.picker(isDisabled).inputNative,
+});
+
+export default pickerStyles;
